refactor proposal artifact router
